Renamed any “name” field to “displayName”
The REST API extensively uses a “name” field for the resource name of the resource. For example the “name” (resource name) of a device would be “projects/<project_id>/devices/<device_id>”. This Swift API parses this resource name into an identifier (device_id) and a parent identifier (project_id). The name property is instead defined as “labels.name” to get the display name as set by a user.

To avoid confusion between these two concepts, this commit removes any property or function with “name” in its name, and replaces it with “displayName”. Identifiers will still be called “identifier”