feat: Adding support for product numbers
Product numbers are now available on DT devices. This shows up as a new `productNumber` field field on the `Device` struct.

Product numbers can be used to filter devices both when listing devices (`getAllDevices` or `getDevicesPage`) and when streaming devices (`subscribeToDevices`). They can also be used to sort devices using the `orderBy` argument.