feat: Added support for 2nd gen temperature sensor events
The 2nd generation DT temperature sensor has a new field in the temperature event called `samples`. This contains temperature measurements made between two heartbeats, allowing for a higher sample rate of temperature measurements.

This adds support for that new `samples` field. The order of the samples is the same as for events, meaning that the oldest event is at the beginning of the array.