Added timestamp to LabelsChangedEvent
This is sort of hacky because labelsChanged events does not contain an `updateTime` in their inner `data` object like the other events. Instead, it has to be read from the outer `timestamp` field, which is not included with the data decoding. This commit adds it to the event struct after decoding the event data, but temporarily sets it to `Date()`.