Centralized authentication
Changed the `Disruptive` struct from being something that should be initialized once and then passed through the whole codebase, to being an encapsulation of global state, including authentication.

This should hopefully make the client library more easy to use in generale Also opens up for having the endpoints on the resources themselves, rather than on the Disruptive struct.