Changed all endpoint to being static functions on the resource structs
This means that instead of passing around an instance of a Disruptive struct, the endpoints can be called as static function on the resource structs themselves. Eg. instead of calling `disruptive.getDevices` on an instance of `Disruptive` that has been passed through multiple function calls, you can just call `Device.getDevices` from anywhere in your code.