Rename endpoints to be more concise
After the endpoints got moved from the Disruptive struct to the resource structs (Device, DataConnector, etc), the function names ended up being very verbose with repeating names.

This commet changes that so that all words in a call conveys salient information. Instead of `DataConnector.createDataConnector(...)` it now becomes `DataConnector.create(...)`.

This is a source breaking change for more or less all the endpoints.